+++ packages/opencode/src/session/message-v2.ts
@@ -331,6 +331,11 @@ export namespace MessageV2 {
     ),
   }
 
+  export type WithParts = {
+    info: Info
+    parts: Part[]
+  }
+
   export function fromV1(v1: Message.Info) {
     if (v1.role === "assistant") {
       const info: Assistant = {
@@ -551,5 +556,11 @@ export namespace MessageV2 {
     }
 
     return convertToModelMessages(result)
+  }
+
+  export function filterSummarized(msgs: { info: MessageV2.Info; parts: MessageV2.Part[] }[]) {
+    const i = msgs.findLastIndex((m) => m.info.role === "assistant" && !!m.info.summary)
+    if (i === -1) return msgs.slice()
+    return msgs.slice(i)
   }
 }
