+++ packages/opencode/src/tool/task.ts
@@ -6,6 +6,7 @@ import { Bus } from "../bus"
 import { MessageV2 } from "../session/message-v2"
 import { Identifier } from "../id/id"
 import { Agent } from "../agent/agent"
+import { SessionPrompt } from "../session/prompt"
 
 export const TaskTool = Tool.define("task", async () => {
   const agents = await Agent.list().then((x) => x.filter((a) => a.mode !== "primary"))
@@ -49,9 +50,9 @@ export const TaskTool = Tool.define("task", async () => {
       }
 
       ctx.abort.addEventListener("abort", () => {
-        Session.abort(session.id)
+        SessionPrompt.abort(session.id)
       })
-      const result = await Session.prompt({
+      const result = await SessionPrompt.prompt({
         messageID,
         sessionID: session.id,
         model: {
